# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface TelegramStatus {
  connected: boolean;
  /** @nullable */
  botUsername: string | null;
  /** @nullable */
  lastUpdateAt: string | null;
  /** @nullable */
  lastError: string | null;
}

export type MemoryKind = typeof MemoryKind[keyof typeof MemoryKind];


export const MemoryKind = {
  message: 'message',
  photo: 'photo',
  reminder: 'reminder',
  debt: 'debt',
  note: 'note',
} as const;

export type MemoryDataDirection = typeof MemoryDataDirection[keyof typeof MemoryDataDirection];


export const MemoryDataDirection = {
  owedToMe: 'owedToMe',
  iOwe: 'iOwe',
} as const;

export interface MemoryData {
  amount?: number;
  direction?: MemoryDataDirection;
  title?: string;
  date?: string;
  fileId?: string;
  caption?: string;
  person?: string;
}

export interface MemoryEntry {
  id: string;
  chatId: string;
  kind: MemoryKind;
  text: string;
  createdAt: string;
  data?: MemoryData;
}

export type AssistantMemorySummaryCounts = {
  /** @minimum 0 */
  reminders: number;
  /** @minimum 0 */
  debts: number;
  /** @minimum 0 */
  notes: number;
  /** @minimum 0 */
  photos: number;
};

export interface AssistantMemorySummary {
  entries: MemoryEntry[];
  counts: AssistantMemorySummaryCounts;
}

export interface CommandRequest {
  /** @minLength 1 */
  text: string;
}

export type GetAssistantMemoryParams = {
/**
 * @minimum 1
 * @maximum 500
 */
limit?: number;
};