# lib/api-zod/src/generated/types/assistantMemorySummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AssistantMemorySummaryCounts } from './assistantMemorySummaryCounts';
import type { MemoryEntry } from './memoryEntry';

export interface AssistantMemorySummary {
  entries: MemoryEntry[];
  counts: AssistantMemorySummaryCounts;
}