# lib/api-zod/src/generated/types/memoryData.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MemoryDataDirection } from './memoryDataDirection';

export interface MemoryData {
  amount?: number;
  direction?: MemoryDataDirection;
  title?: string;
  date?: string;
  fileId?: string;
  caption?: string;
  person?: string;
}