# lib/api-zod/src/generated/types/memoryEntry.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MemoryData } from './memoryData';
import type { MemoryKind } from './memoryKind';

export interface MemoryEntry {
  id: string;
  chatId: string;
  kind: MemoryKind;
  text: string;
  createdAt: Date;
  data?: MemoryData;
}