# lib/api-zod/src/generated/types/memoryKind.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type MemoryKind = typeof MemoryKind[keyof typeof MemoryKind];


export const MemoryKind = {
  message: 'message',
  photo: 'photo',
  reminder: 'reminder',
  debt: 'debt',
  note: 'note',
} as const;