# lib/api-zod/src/generated/types/telegramStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TelegramStatus {
  connected: boolean;
  /** @nullable */
  botUsername: string | null;
  /** @nullable */
  lastUpdateAt: Date | null;
  /** @nullable */
  lastError: string | null;
}